Fix AggregateFunctionsTests#stddev
The previous epsilon based approximatelyEqual test failed with large number. Let's take an example to illustrate: suppose Spark finds a stddev of 1.768294153521034E78 and list computed version founds 1.7682941535210339E78. These two are pretty close, yet their diff is "huge": ~2E62. This means that this diff should also be compared to a fraction of either stddev in addition to the fixed epsilon.