More macros!
Macro all the things!

A new macro for `selectExpr[B](T => B)`. It's like `map`, only it only
works if the provided function can be converted to a projection of some
column expressions over the original dataset - so it avoids SerDe.

It won't work all the time, but (hopefully) gives you good compiler
errors when it doesn't work.

Currently it drops down into `DataFrame` to create the projection plan,
and unifies it back into `TypedDataset` (via `TypedDataset.create` over
`Dataset`).

There are a couple of hairy pieces that could be cleaned up (i.e. it's
not clear what to do about SQL functions and UDFs, since they don't
operate on the actual type). But it's already at a point where it's
pretty useful.