Added functions to selectExpr
`selectExpr` now supports functions. The scheme I've come up with has a
mirror of each Spark column function, but with a type signature that
describes the actual arguments and result type of the function. These
mirrors (in frameless.functions.quoted) are annotated with a variant of
QuotedFunc, which specifies which Spark function they should rewrite to.
The macro picks up these annotations and allows the function call to be
rewritten to the native spark function once the column expressions are
refolded.

Roughly half of functions are mirrored at this point; need to finish
porting aggregate and misc functions.