Add `lit`
Had to do the same trick as in `FramelessUdf` because our encoders don't
support `eval`.

As in Spark, doesn't work unless value is Serializable. If our encoders
supported `eval`, we could value before distributing it among cluster
nodes. It's hardly doable right now, not sure if it worth spending time.

It's open question if we should implement `eval` in `TypedEncoder` or
not, at the moment it's still not required.