Remove TypedColumn's first type parameter
This change reduces the type saftely of column expressions. For instance, it's now possible to typecheck some nonsensical `select` as follows:

```
ds1.select(ds2('a)) // This use to be a type error!
```

This simplification is a required step to support joins, which is also source compatible. Safety can hopefuly be recovered by having `select` take a `TypedColumn[T] => TypedColumn[C]` function and removing the apply method on TypedDataset.