Anonymize/reformat all implicit parameter lists
I don't think the names of implicits help at all for readability. To
understand what a shapeless program does it's best to look at the type! Also,
because the type arguments are often the most important part of a function
definition it makes sense to have them in their own level of indentation as
implemented in this PR.