And col{Left,Right} and Disambiguate{Left,Right}
This commit introduces a disambiguation mechanism for self-joins. The idea is
to use ds.colLeft('a) and ds.colRight('a) instead of ds.col('a) when writing
join expression, where the Left and Right suffix are use to tag the column for
being on the left hand side of on the right hand side of the join expression.
The implementation adds two new nodes to the Spark IR DisambiguateLeft and
DisambiguateRight, which behave at dummies (just forward to underlying) but
are intersepted at runtime when building the join plans to actually implement
the disambiguation.