Joins with boolean column expressions, 2nd attempt (#175)
* Minor changes

* Add With type class

With computes the intersection of two types while leaving less noise than A with B:

- With[A, A] = A
- With[A, B] = A with B (when A != B)

This type function is needed to prevent IDEs from infering large types
with shape `A with A with ... with A`. These types could be confusing for
both end users and IDE's type checkers.

* Use With in all TypedColumn methods

* Remove automagic widden import

Implicit conversions don't compose; I propose we simply cast everything explicitly.

* And col{Left,Right} and Disambiguate{Left,Right}

This commit introduces a disambiguation mechanism for self-joins. The idea is
to use ds.colLeft('a) and ds.colRight('a) instead of ds.col('a) when writing
join expression, where the Left and Right suffix are use to tag the column for
being on the left hand side of on the right hand side of the join expression.
The implementation adds two new nodes to the Spark IR DisambiguateLeft and
DisambiguateRight, which behave at dummies (just forward to underlying) but
are intersepted at runtime when building the join plans to actually implement
the disambiguation.

* Add joins APIs

* Test joins